requirements.txt
torch
torchvision
tqdm
Pillow
numpy
matplotlib
opencv-python
opencv-python-contrib
# jupyterlab
# ipywidgets
# ipykernel
